resolve error in creating gps table
When extracted TDOA tracks did not start at the beginning of the encounters, there was an error in creating a gps table. The issue was that beginning and end times for searching the bigger database were based on detected tracks. Instead should have been based on an overall time vector `t_serialdate`.